refactor: 폴더 구조 정리 및 그라운드 상세 조회
