+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Cookies from 'js-cookie';
-import WithTokenRouter from './pages/with-token/WithTokenRouter';
-import WithoutTokenRouter from './pages/without-token/WithoutTokenRouter';
+import WithoutTokenRouter from './router/WithoutTokenRouter';
+import WithTokenRouter from './router/WithTokenRouter';
 
 function App() {
   const token = Cookies.get('Authorization');
