+++ src/components/GroundList.jsx
@@ -1,12 +1,9 @@
 import React from 'react';
-import GroundCard from './GroundCard';
 
 function GroundList() {
   return (
     <div>
-      <GroundCard />
-      <GroundCard />
-      <GroundCard />
+      list
     </div>
   );
 }
