+++ src/pages/with-token/Home.jsx
@@ -9,6 +9,7 @@ function Home() {
     <Layout>
       <a href="/login">로그인</a>
       <a href="/register">회원가입</a>
+      <a href="/profile">프로필</a>
       <CurrentGround />
       <Search />
       <GroundList />
