+++ src/pages/with-token/Profile.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import Layout from '../../components/Layout';
 
-function CreateGroundPage() {
+function Profile() {
   return (
     <Layout>
-      ground form 넣기
+      프로필!
     </Layout>
   );
 }
-export default CreateGroundPage;
+export default Profile;
